Small refactor to simplify the code.

+++ ATM90E32.h
@@ -398,7 +398,7 @@ public:
 	double GetTemperature();
 
 	/* Gain Parameters (GET)*/
-	double GetValueRegister(unsigned short registerRead);
+	unsigned short GetValueRegister(unsigned short registerRead);
 
 	/* Energy Consumption */
 	double GetImportEnergy();
